% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topomatch.R
\name{best_matches}
\alias{best_matches}
\title{Best matches}
\usage{
best_matches(x)
}
\arguments{
\item{x}{\code{topomatch} object}
}
\value{
The original list of toponyms with their best-ranked
  candidates.
}
\description{
Best matches
}
\examples{
  best_matches(topomatch(c("a", "b"), c("A", "X")))
}
