% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topomatch.R
\name{exact_matches}
\alias{exact_matches}
\title{Exact toponym matches}
\usage{
exact_matches(x)
}
\arguments{
\item{x}{\code{topomatch} object}
}
\value{
named logical vector
}
\description{
Exact toponym matches
}
\examples{
  exact_matches(topomatch(c("a", "b"), "A"))
}
