% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topomatch.R
\name{similar_matches}
\alias{similar_matches}
\title{Similar toponym matches}
\usage{
similar_matches(x)
}
\arguments{
\item{x}{\code{topomatch} object}
}
\value{
named logical vector
}
\description{
Similar toponym matches
}
\examples{
  similar_matches(topomatch(c("a", "b"), "A"))
}
