% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topomatch.R
\name{topomatch}
\alias{topomatch}
\title{Toponomy matching}
\usage{
topomatch(x, candidates)
}
\arguments{
\item{x}{character vector of terms to matcih}

\item{candidates}{character vector of terms to compare against}
}
\value{
A matrix with x in rows and candidates in columns with matching
  scores.
}
\description{
Match a list of toponyms agains a list of candidates.
}
\details{
For comparison, strings are trimmed off whitespaces, converted to
upper-case and special characters transliterated to ASCII.
}
\examples{
  x <- c("abá", "Efé", "IJI", "kkk")
  cand <- c("aba", "ef", "iji", "opo", "Uvu")
  topomatch(x, cand)
}
