% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcribe.R
\name{transcribe}
\alias{transcribe}
\title{Transcribe toponyms according to a topomatch table}
\usage{
transcribe(obj, fixes = NULL)
}
\arguments{
\item{obj}{A \code{topomatch} object.}

\item{fixes}{A Named character vector of manual fixes.}
}
\value{
A function that takes a character vector among the original
terms and returns a character vector with the corresponding matches.
}
\description{
Build a function for transcription.
}
\details{
If there are any unmatched toponyms, they must be manually matched
in fixes for transcription. On the other hand, fixes can provide matches for incorrect
associations. See examples.
}
\examples{
  (perfect <- topomatch("AFGHANISTAN", "Afghanistan"))
  transcribe(perfect)("Afghanistan")

  terms <- c("foo", "bar", "foobar", "X")
  (mat <- topomatch(terms, c("Foo", "Bar")))
  setNames(transcribe(mat, fixes = c(X = "Bar"))(terms), terms)

  ## You can still transcribe terms that do not need fixes
  # transcribe(mat)(terms)  # Error: X is not fixed
  transcribe(mat)(c("foo", "bar"))  # Works even if X remains unfixed.

}
