% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.R
\name{trim}
\alias{trim}
\title{Trim whitespaces}
\usage{
trim(x, internal = TRUE)
}
\arguments{
\item{x}{character vector, matrix or array.}

\item{internal}{logical. Whether to replace sequential internal by
a single space}
}
\value{
A character object of the same type.
}
\description{
Remove the whitespace before or after a string of character
}
\examples{
  trim(c("    hi folks    !   ", "    hello   world    "))
}
\author{
Robert J. Hijmans and Jacob van Etten. The code was borrowed
from function \code{\link[raster]{trim}} in package \code{raster}.
}
