% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topomatch.R
\name{unmatched}
\alias{unmatched}
\title{Unmatched toponyms}
\usage{
unmatched(x)
}
\arguments{
\item{x}{\code{topomatch} object}
}
\value{
named logical vector
}
\description{
Vector of toponyms with more than one plausible candidate with
the same score.
}
\examples{
  unmatched(topomatch(c("a", "b"), c("A", "X")))
}
